Add zeros, ones, fix compiler warnings (#3679)
* Add zeros, ones, fix compiler warnings

* Move things to underscore, fix compile warning in scala pkg

+++ scala-package/native/src/main/native/ml_dmlc_mxnet_native_c_api.cc
@@ -167,7 +167,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxImperativeInvoke
   if (numParams > 0) {
     cParamKeys = new const char *[numParams];
     cParamVals = new const char *[numParams];
-    for (size_t i = 0; i < numParams; i++) {
+    for (int i = 0; i < numParams; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(paramKeys, i));
       const char *key = env->GetStringUTFChars(jkey, 0);
       cParamKeys[i] = key;
@@ -204,7 +204,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxImperativeInvoke
 
   // release allocated memory
   if (numParams > 0) {
-    for (size_t i = 0; i < numParams; i++) {
+    for (int i = 0; i < numParams; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(paramKeys, i));
       env->ReleaseStringUTFChars(jkey, cParamKeys[i]);
       env->DeleteLocalRef(jkey);
@@ -261,7 +261,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxFuncInvokeEx
   if (numParams > 0) {
     cParamKeys = new jbyte *[numParams];
     cParamVals = new jbyte *[numParams];
-    for (size_t i = 0; i < numParams; i++) {
+    for (int i = 0; i < numParams; i++) {
       jbyteArray jkey = reinterpret_cast<jbyteArray>(env->GetObjectArrayElement(paramKeys, i));
       jbyte *cParamKey = env->GetByteArrayElements(jkey, NULL);
       cParamKeys[i] = cParamKey;
@@ -283,7 +283,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxFuncInvokeEx
   env->ReleaseFloatArrayElements(scalarArgs, cScalarArgs, 0);
   env->ReleaseLongArrayElements(mutateVars, cMutateVars, 0);
   if (numParams > 0) {
-    for (size_t i = 0; i < numParams; i++) {
+    for (int i = 0; i < numParams; i++) {
       jbyteArray jkey = reinterpret_cast<jbyteArray>(env->GetObjectArrayElement(paramKeys, i));
       env->ReleaseByteArrayElements(jkey, cParamKeys[i], 0);
       env->DeleteLocalRef(jkey);
@@ -470,7 +470,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxNDArraySave
   const char **keys = NULL;
   if (jkeys != NULL) {
     keys = new const char *[numArgs];
-    for (size_t i = 0; i < numArgs; i++) {
+    for (int i = 0; i < numArgs; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
       const char *key = env->GetStringUTFChars(jkey, 0);
       keys[i] = key;
@@ -489,7 +489,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxNDArraySave
 
   // release allocated memory
   if (jkeys != NULL) {
-    for (size_t i = 0; i < numArgs; i++) {
+    for (int i = 0; i < numArgs; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
       env->ReleaseStringUTFChars(jkey, keys[i]);
       env->DeleteLocalRef(jkey);
@@ -516,7 +516,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxInitPSEnv
   const char** vals = new const char*[paramSize];
   jstring jkey, jval;
   // use strcpy and release char* created by JNI inplace
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
     const char* ckey = env->GetStringUTFChars(jkey, 0);
     keys[i] = ckey;
@@ -533,7 +533,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxInitPSEnv
                         static_cast<const char**>(vals));
 
   // release keys and vals
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jstring key = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
     env->ReleaseStringUTFChars(key, keys[i]);
     env->DeleteLocalRef(key);
@@ -858,7 +858,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxDataIterCreateIter
   const char** vals = new const char*[paramSize];
   jstring jkey, jval;
   // use strcpy and release char* created by JNI inplace
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
     const char* ckey = env->GetStringUTFChars(jkey, 0);
     keys[i] = ckey;
@@ -880,7 +880,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxDataIterCreateIter
   SetLongField(env, dataIterHandleRef, reinterpret_cast<jlong>(out));
 
   // release keys and vals
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jstring key = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
     env->ReleaseStringUTFChars(key, keys[i]);
     env->DeleteLocalRef(key);
@@ -1069,7 +1069,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolCreateAtomicSymbol
   int paramSize = env->GetArrayLength(paramKeys);
   const char **keys = new const char*[paramSize];
   const char **vals = new const char*[paramSize];
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jstring key = reinterpret_cast<jstring>(env->GetObjectArrayElement(paramKeys, i));
     const char *rawKey = env->GetStringUTFChars(key, 0);
     keys[i] = rawKey;
@@ -1087,7 +1087,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolCreateAtomicSymbol
   SetLongField(env, symbolRef, reinterpret_cast<jlong>(out));
 
   // release keys and vals
-  for (size_t i = 0; i < paramSize; i++) {
+  for (int i = 0; i < paramSize; i++) {
     jstring key = reinterpret_cast<jstring>(env->GetObjectArrayElement(paramKeys, i));
     env->ReleaseStringUTFChars(key, keys[i]);
     env->DeleteLocalRef(key);
@@ -1119,7 +1119,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolCompose
   const char **keys = NULL;
   if (jkeys != NULL) {
     keys = new const char*[argSize];
-    for (size_t i = 0; i < argSize; i++) {
+    for (int i = 0; i < argSize; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
       const char *key = env->GetStringUTFChars(jkey, 0);
       keys[i] = key;
@@ -1135,7 +1135,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolCompose
   env->ReleaseLongArrayElements(jargs, args, 0);
   // release allocated memory
   if (jkeys != NULL) {
-    for (size_t i = 0; i < argSize; i++) {
+    for (int i = 0; i < argSize; i++) {
       jstring jkey = (jstring) env->GetObjectArrayElement(jkeys, i);
       env->ReleaseStringUTFChars(jkey, keys[i]);
       env->DeleteLocalRef(jkey);
@@ -1275,7 +1275,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolInferType
   const char **keys = NULL;
   if (jkeys != NULL) {
     keys = new const char *[numArgs];
-    for (size_t i = 0; i < numArgs; i++) {
+    for (int i = 0; i < numArgs; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
       const char *key = env->GetStringUTFChars(jkey, 0);
       keys[i] = key;
@@ -1462,7 +1462,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxSymbolInferShape
 
   // release allocated memory
   if (jkeys != NULL) {
-    for (size_t i = 0; i < jnumArgs; i++) {
+    for (int i = 0; i < jnumArgs; i++) {
       jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jkeys, i));
       env->ReleaseStringUTFChars(jkey, keys[i]);
       env->DeleteLocalRef(jkey);
@@ -1482,7 +1482,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxExecutorBindX
   int auxStatesLen = env->GetArrayLength(jauxArgsHandle);
 
   const char **mapKeys = new const char *[numCtx];
-  for (size_t i = 0; i < numCtx; i++) {
+  for (int i = 0; i < numCtx; i++) {
     jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jctxMapKeys, i));
     const char *key = env->GetStringUTFChars(jkey, 0);
     mapKeys[i] = key;
@@ -1514,7 +1514,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxExecutorBindX
   env->ReleaseLongArrayElements(jargsHandle, inArgs, 0);
   env->ReleaseIntArrayElements(jreqsArray, gradReqType, 0);
   env->ReleaseLongArrayElements(jauxArgsHandle, auxStates, 0);
-  for (size_t i = 0; i < numCtx; i++) {
+  for (int i = 0; i < numCtx; i++) {
     jstring jkey = (jstring) env->GetObjectArrayElement(jctxMapKeys, i);
     env->ReleaseStringUTFChars(jkey, mapKeys[i]);
     env->DeleteLocalRef(jkey);
@@ -1536,7 +1536,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxExecutorBindEX
   if ((int32_t)jsharedExec != 0) sharedExec = reinterpret_cast<ExecutorHandle>(jsharedExec);
 
   const char **mapKeys = new const char *[numCtx];
-  for (size_t i = 0; i < numCtx; i++) {
+  for (int i = 0; i < numCtx; i++) {
     jstring jkey = reinterpret_cast<jstring>(env->GetObjectArrayElement(jctxMapKeys, i));
     const char *key = env->GetStringUTFChars(jkey, 0);
     mapKeys[i] = key;
@@ -1569,7 +1569,7 @@ JNIEXPORT jint JNICALL Java_ml_dmlc_mxnet_LibInfo_mxExecutorBindEX
   env->ReleaseLongArrayElements(jargsHandle, inArgs, 0);
   env->ReleaseIntArrayElements(jreqsArray, gradReqType, 0);
   env->ReleaseLongArrayElements(jauxArgsHandle, auxStates, 0);
-  for (size_t i = 0; i < numCtx; i++) {
+  for (int i = 0; i < numCtx; i++) {
     jstring jkey = (jstring) env->GetObjectArrayElement(jctxMapKeys, i);
     env->ReleaseStringUTFChars(jkey, mapKeys[i]);
     env->DeleteLocalRef(jkey);
